lab4_4/pyproject.toml
[project]
name = "lab4_4"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Atr1ck", email = "3126670240@qq.com"},
]
dependencies = ["antlr4-python3-runtime>=4.13.2"]
requires-python = "==3.8.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
